python_src/requirements.txt
pygame==2.5.2
pyserial==3.5
scikit-learn==1.3.2
numpy==1.26.4
joblib==1.3.2
